improve OO, db OK
