+++ src/backend/classes/__init__.py
@@ -1,1 +1,1 @@
-print('Running file: __init__.py from app/classes/extrator')
+#print('Running file: __init__.py from app/classes/extrator')
