+++ src/backend/classes/city/__init__.py
@@ -1,1 +1,1 @@
-print('Running file: __init__.py from app/classes/city')
+#print('Running file: __init__.py from app/classes/city')
